feat(core): finish Analysis ci
