+++ .gitignore
@@ -2,6 +2,7 @@ node_modules
 
 .vscode
 .DS_Store
+.intercepter
 
 old
 dist
