+++ package.json
@@ -11,7 +11,9 @@
   "scripts": {
     "build:prepare": "pnpm run --filter @blog/cli... build",
     "build": "blog build --outDir=dist --mode=production",
+    "build:debug": "pnpm run build --debug",
     "watch": "blog watch --hmr --mode=development",
+    "watch:debug": "pnpm run watch --debug",
     "test": "pnpm run -r --parallel test",
     "lint": "pnpm run -r --parallel lint --no-error-on-unmatched-pattern",
     "lint:fix": "pnpm run -r --parallel lint:fix --no-error-on-unmatched-pattern"
