+++ packages/builder/cli/bin/index.js
@@ -1,1 +1,2 @@
+require('dotenv').config();
 require('../dist').run();
