+++ packages/builder/cli/package.json
@@ -21,14 +21,15 @@
   "dependencies": {
     "@blog/core": "workspace:*",
     "@blog/types": "workspace:*",
+    "dotenv": "16.0.3",
     "tslib": "2.4.1",
     "yargs": "17.5.1"
   },
   "devDependencies": {
-    "@xiao-ai/eslint-config": "1.0.1",
-    "@xiao-ai/tsconfig": "1.0.1",
     "@types/node": "16.11.7",
     "@types/yargs": "17.0.11",
+    "@xiao-ai/eslint-config": "1.0.1",
+    "@xiao-ai/tsconfig": "1.0.1",
     "eslint": "8.31.0",
     "rimraf": "3.0.2",
     "typescript": "4.9.4"
