+++ packages/builder/core/src/plugins/intercepter/hooks/constant.ts
@@ -1,0 +1,4 @@
+import type { AllPluginData } from './types';
+
+/** 数据储存 */
+export const hookData: AllPluginData = new Map();
