+++ packages/builder/core/src/plugins/intercepter/index.ts
@@ -1,4 +1,6 @@
 import type { BuilderPlugin } from '@blog/types';
+import { join } from 'path';
+import { writeFile } from '@blog/node';
 import type { DebuggerOptions } from './types';
 
 import * as Hook from './hooks';
@@ -13,20 +15,24 @@ export const Intercepter = (opt?: DebuggerOptions): BuilderPlugin => {
         return;
       }
 
-      const { logger } = builder;
-      const options: Required<DebuggerOptions> = {
-        // 默认排除自己
+      const { logger, options } = builder;
+      const pluginOptions: DebuggerOptions = {
         excludes: [pluginName].concat(opt?.excludes ?? []),
+        outFile: opt?.outFile ? join(options.root, opt?.outFile) : undefined,
       };
 
-      Hook.intercept(pluginName, builder, options);
+      Hook.intercept(pluginName, builder, pluginOptions);
 
       builder.hooks.start.tap(pluginName, () => {
         Hook.clear();
       });
 
-      builder.hooks.success.tap(pluginName, () => {
-        Hook.printHookData(logger);
+      builder.hooks.success.tapPromise(pluginName, async () => {
+        Hook.print(logger);
+
+        if (pluginOptions.outFile) {
+          await writeFile(pluginOptions.outFile, Hook.getMdString());
+        }
       });
     },
   };
