+++ packages/utils/node/src/index.ts
@@ -1,2 +1,3 @@
 export * from './string';
 export * from './path';
+export * from './file';
